requirements.txt
discord.py
python-dotenv~=0.21.0

discord~=2.0.0
pip~=20.0.2
keyring~=18.0.1
idna~=2.8
multidict~=6.0.2
attrs~=19.3.0
yarl~=1.8.1
aiohttp~=3.8.3
aiosignal~=1.2.0
frozenlist~=1.3.1
requests~=2.22.0
setuptools~=45.2.0